:wastebasket: temp: 추천 모달 처리
